solution/requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.9.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bitsandbytes==0.41.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickhouse-connect==0.7.7
clickhouse-driver==0.2.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
einops==0.6.1
einops-exts==0.0.4
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gradio==3.35.2
gradio_client==0.2.9
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
libretranslatepy==2.1.1
linkify-it-py==2.0.3
lxml==5.2.1
lz4==4.3.3
markdown-it-py==2.2.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multilingual-clip==1.0.10
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
open-clip-torch==2.24.0
orjson==3.10.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathos==0.3.2
peft==0.4.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pox==0.3.4
ppft==1.7.6.8
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.1
pydub==0.25.1
PyExecJS==1.5.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.7
safetensors==0.4.2
scikit-learn==1.2.2
scipy==1.13.0
semantic-version==2.10.0
sentence-transformers==2.6.1
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
threadpoolctl==3.4.0
tiktoken==0.6.0
timm==0.6.13
-e git+https://github.com/DLCV-BUAA/TinyLLaVABench.git@2d2105fc7d44781464e7f503986bbfac270a8e22#egg=tinyllava
tokenizers==0.15.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.37.2
translate==3.6.1
translators==5.9.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.29.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
yarl==1.9.4
zstandard==0.22.0
